Move Vector to utils module
