+++ platformer.py
@@ -10,33 +10,7 @@ import pygame.time
 from pygame import Color, Surface
 from pygame.sprite import Group, Sprite
 
-
-class Vector(namedtuple('Vector', ['x', 'y'])):
-    def __neg__(self):
-        return Vector(-self.x, -self.y)
-
-    def __add__(self, other):
-        if not isinstance(other, Vector):
-            raise TypeError(f"{other} is not a vector")
-
-        x = self.x + other.x
-        y = self.y + other.y
-        return Vector(x, y)
-
-    def __sub__(self, other):
-        return self.__add__(-other)
-
-    def __mul__(self, other):
-        if isinstance(other, Number):
-            return Vector(other * self.x, other * self.y)
-        else:
-            raise TypeError(f"can't multiply vector by {type(other)}")
-
-    def __rmul__(self, other):
-        return self.__mul__(other)
-
-    def __truediv__(self, other):
-        return self.__mul__(1 / other)
+from utils import Vector
 
 
 FPS = 60
